// frontend/@types/api/models/CreateAvailabilityDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SkillSwap API
 * API for the SkillSwap skills exchange platform
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface CreateAvailabilityDto
 */
export interface CreateAvailabilityDto {
    /**
     * The day of the week (0-6, where 0 is Sunday and 6 is Saturday)
     * @type {number}
     * @memberof CreateAvailabilityDto
     */
    day: number;
    /**
     * Start time of the availability
     * @type {Date}
     * @memberof CreateAvailabilityDto
     */
    startTime: Date;
    /**
     * End time of the availability
     * @type {Date}
     * @memberof CreateAvailabilityDto
     */
    endTime: Date;
    /**
     * Unique identifier for the user
     * @type {string}
     * @memberof CreateAvailabilityDto
     */
    userId: string;
}

/**
 * Check if a given object implements the CreateAvailabilityDto interface.
 */
export function instanceOfCreateAvailabilityDto(
    value: object
): value is CreateAvailabilityDto {
    if (!("day" in value) || value["day"] === undefined) return false;
    if (!("startTime" in value) || value["startTime"] === undefined)
        return false;
    if (!("endTime" in value) || value["endTime"] === undefined) return false;
    if (!("userId" in value) || value["userId"] === undefined) return false;
    return true;
}

export function CreateAvailabilityDtoFromJSON(
    json: any
): CreateAvailabilityDto {
    return CreateAvailabilityDtoFromJSONTyped(json, false);
}

export function CreateAvailabilityDtoFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean
): CreateAvailabilityDto {
    if (json == null) {
        return json;
    }
    return {
        day: json["day"],
        startTime: new Date(json["startTime"]),
        endTime: new Date(json["endTime"]),
        userId: json["userId"],
    };
}

export function CreateAvailabilityDtoToJSON(json: any): CreateAvailabilityDto {
    return CreateAvailabilityDtoToJSONTyped(json, false);
}

export function CreateAvailabilityDtoToJSONTyped(
    value?: CreateAvailabilityDto | null,
    ignoreDiscriminator: boolean = false
): any {
    if (value == null) {
        return value;
    }

    return {
        day: value["day"],
        startTime: value["startTime"].toISOString(),
        endTime: value["endTime"].toISOString(),
        userId: value["userId"],
    };
}